[Misc] Add attention mask
