+++ vllm_ascend/attention/attention.py
@@ -35,6 +35,7 @@ from vllm.attention.backends.utils import (PAD_SLOT_ID, CommonAttentionState,
 from vllm.utils import async_tensor_h2d, make_tensor_with_pad
 
 from vllm_ascend.ascend_config import get_ascend_config
+from vllm_ascend.attention.attention_mask import AttentionMaskBuilder
 from vllm_ascend.ops.cache import concat_and_cache_mla
 from vllm_ascend.utils import (ACL_FORMAT_FRACTAL_NZ, aligned_16,
                                enable_custom_op, is_310p, nd_to_nz_2d)
@@ -42,108 +43,6 @@ from vllm_ascend.worker.model_runner import (
     ModelInputForNPUBuilder, ModelInputForNPUWithSamplingMetadata)
 
 _ALLOWED_NUM_QUERIES_PER_KV = [32, 64, 128]
-
-
-def generate_attn_mask(max_seq_len: int, dtype=torch.float16, mask_value=None):
-    # Construct lower triangle matrix.
-    mask_flag = torch.tril(
-        torch.ones((max_seq_len, max_seq_len),
-                   dtype=torch.bool)).view(max_seq_len, max_seq_len)
-    # Create upper triangle matrix used to mark mask positions.
-    mask_flag = ~mask_flag
-    # Currently for fp16 dtype, the mask value should be set to -inf.
-    # TODO: Eliminate this part in the future.
-    if mask_value is None:
-        if dtype == torch.float16:
-            mask_value = torch.finfo(torch.float32).min
-        else:
-            mask_value = 1
-    attn_mask = torch.masked_fill(torch.zeros(size=(max_seq_len, max_seq_len)),
-                                  mask_flag, mask_value).to(dtype)
-    return attn_mask
-
-
-class AttentionMaskBuilder:
-
-    def __init__(self, attn_mask: torch.Tensor):
-        self._seq_len_cached = attn_mask.shape[0]
-        self.attn_mask_cache = attn_mask
-        self.splitfuse_mask_value = -10000
-
-    @classmethod
-    def initialize_from_len(cls,
-                            max_seq_len: int,
-                            dtype: torch.dtype = torch.float16,
-                            mask_value: Optional[int] = None):
-        return cls(generate_attn_mask(max_seq_len, dtype, mask_value))
-
-    def update_attn_cache(self, seqlen: int, dtype: torch.dtype,
-                          device: torch.device):
-        if seqlen > self._seq_len_cached or self.attn_mask_cache.dtype != dtype:
-            self._seq_len_cached = seqlen
-            self.attn_mask_cache = generate_attn_mask(seqlen, dtype)
-        if self.attn_mask_cache.device != device:
-            self.attn_mask_cache = self.attn_mask_cache.to(device)
-
-    def get_attn_mask(self, max_seq_len: int, dtype: torch.dtype,
-                      device: torch.device):
-        self.update_attn_cache(max_seq_len, dtype, device)
-        return self.attn_mask_cache[:max_seq_len, :max_seq_len].contiguous()
-
-    def get_decode_attn_mask(
-        self,
-        input_lengths: torch.tensor,
-        max_s: int,
-        dtype: torch.dtype,
-        device: torch.device,
-    ):
-        self.update_attn_cache(max_s, dtype, device)
-        return (self.attn_mask_cache.index_select(
-            0, input_lengths)[:, :max_s].view(-1, 1, max_s).contiguous())
-
-    def get_splitfuse_attn_mask(
-        self,
-        seq_lens,
-        query_lens,
-        position,
-        dtype,
-        device,
-    ) -> torch.Tensor:
-        max_seq_len = max(seq_lens, default=0)
-        if max_seq_len <= self._seq_len_cached:
-            self.update_attn_cache(max_seq_len, dtype, device)
-            # FIXME: Currently the mask value of chunked-prefill situation and Prefill-Only situation
-            # is not the same. Fix this in the future when kernel is ready.
-            if self.attn_mask_cache.numel(
-            ) > 1 and self.attn_mask_cache[0][1] > 0:
-                attn_mask = self.get_attn_mask(  # type: ignore
-                    max_seq_len, dtype, device)
-                attn_mask *= -10000
-            else:
-                attn_mask = self.attn_mask_cache
-            return torch.index_select(attn_mask, dim=0,
-                                      index=position)[:, :max_seq_len]
-        total_q_len = sum(query_lens)
-        attn_mask = torch.zeros((total_q_len, max_seq_len),
-                                dtype=dtype,
-                                device="cpu")
-
-        current_row = 0
-        for i in range(len(query_lens)):
-            seq_len = seq_lens[i]
-            q_len = query_lens[i]
-            context_len = seq_len - q_len
-
-            assert context_len >= 0
-            attn_mask[current_row:current_row + q_len,
-                      context_len:] = self.splitfuse_mask_value
-            right_tensor = attn_mask[current_row:current_row + q_len,
-                                     context_len:seq_len]
-            right_tensor.masked_fill_(
-                right_tensor.tril() == self.splitfuse_mask_value, 0)
-            current_row += q_len
-
-        return attn_mask.to(device, non_blocking=True)
 
 
 class AscendAttentionBackend(AttentionBackend):
@@ -524,7 +423,7 @@ class AscendMetadataBuilder(CommonMetadataBuilder[AscendMetadata]):
         self.compress_mask = None
         self.chunk_mask = None
         if AscendMetadataBuilder._attn_mask_builder is None:
-            AscendMetadataBuilder._attn_mask_builder = AttentionMaskBuilder.initialize_from_len(
+            AscendMetadataBuilder._attn_mask_builder = AttentionMaskBuilder(
                 128, self.input_builder.runner.model_config.dtype)
 
     def _add_seq_group(
